feat: 질문 업데이트, 역할별 컴포넌트 분리
