+++ src/pages/AccountSetting/AccountSetting.tsx
@@ -1,8 +1,16 @@
-import React from 'react';
-import AccountSettingContainer from './components/AccountSettingContainer';
+import { StyledAccountSettingWrap } from './components/AccountSettingUI.styled';
+import PasswordChangeContainer from './components/PasswordChangeContainer';
+import SecurityQuestionContainer from './components/SecurityQuestionContainer';
 
 const AccountSetting = () => {
-  return <AccountSettingContainer />;
+  return (
+    <>
+      <StyledAccountSettingWrap>
+        <PasswordChangeContainer />
+        <SecurityQuestionContainer />
+      </StyledAccountSettingWrap>
+    </>
+  );
 };
 
 export default AccountSetting;
